+++ src/FeatureToggles/Configuration/AppConfig/RoleElement.cs
@@ -1,4 +1,4 @@
-﻿namespace FeatureToggles.Configuration
+﻿namespace FeatureToggles.Configuration.AppConfig
 {
     using System.Configuration;
