+++ src/FeatureToggles/Configuration/AppConfig/ToggleElement.cs
@@ -1,4 +1,4 @@
-﻿namespace FeatureToggles.Configuration
+﻿namespace FeatureToggles.Configuration.AppConfig
 {
     using System.Configuration;
