+++ src/FeatureToggles/Configuration/AppConfig/ToggleElementCollection.cs
@@ -1,19 +1,19 @@
-﻿namespace FeatureToggles.Configuration
+﻿namespace FeatureToggles.Configuration.AppConfig
 {
     using System.Collections.Generic;
     using System.Configuration;
 
-    [ConfigurationCollection(typeof(UsersElement), AddItemName = "user", CollectionType = ConfigurationElementCollectionType.BasicMap)]
-    public class UsersElementCollection : ConfigurationElementCollection, IEnumerable<UsersElement>
+    [ConfigurationCollection(typeof(ToggleElement), AddItemName = "toggle", CollectionType = ConfigurationElementCollectionType.BasicMap)]
+    public class ToggleElementCollection : ConfigurationElementCollection, IEnumerable<ToggleElement>
     {
         /// <summary>
         /// Gets or sets a url element from the collection by index
         /// </summary>
         /// <param name="index">The index</param>
         /// <returns>The url element</returns>
-        public UsersElement this[int index]
+        public ToggleElement this[int index]
         {
-            get => BaseGet(index) as UsersElement;
+            get => BaseGet(index) as ToggleElement;
             set
             {
                 if (Count <= 0)
@@ -35,12 +35,12 @@
         /// Iterator for returning url elements from the collection
         /// </summary>
         /// <returns>A url element</returns>
-        public new IEnumerator<UsersElement> GetEnumerator()
+        public new IEnumerator<ToggleElement> GetEnumerator()
         {
             int count = Count;
             for (int i = 0; i < count; i++)
             {
-                yield return BaseGet(i) as UsersElement;
+                yield return BaseGet(i) as ToggleElement;
             }
         }
 
@@ -51,7 +51,7 @@
         /// <returns>The configuration element</returns>
         protected override ConfigurationElement CreateNewElement()
         {
-            return new UsersElement();
+            return new ToggleElement();
         }
 
         /// <summary>
@@ -61,7 +61,7 @@
         /// <returns>The actual instance required</returns>
         protected override object GetElementKey(ConfigurationElement element)
         {
-            return ((UsersElement)element).Name;
+            return ((ToggleElement)element).Name;
         }
     }
 }
