+++ src/FeatureToggles/Configuration/AppConfig/UsersElement.cs
@@ -1,4 +1,4 @@
-﻿namespace FeatureToggles.Configuration
+﻿namespace FeatureToggles.Configuration.AppConfig
 {
     using System.Configuration;
