+++ src/FeatureToggles/FeatureToggles.csproj
@@ -8,4 +8,9 @@
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.7.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Configuration\AppSettings\Providers\" />
+    <Folder Include="Providers\AppSettings\" />
+  </ItemGroup>
+
 </Project>
