+++ src/FeatureTogglesCoreTests/ConfigurationTests.cs
@@ -22,6 +22,7 @@ namespace ToggleTests
     using System.Configuration;
     using System.Diagnostics.CodeAnalysis;
     using FeatureToggles.Configuration;
+    using FeatureToggles.Configuration.AppConfig;
     using NUnit.Framework;
 
     [TestFixture]
