+++ src/FeatureTogglesFrameworkTests/ToggleTests.cs
@@ -22,6 +22,7 @@ namespace ToggleTests
     using System.Diagnostics.CodeAnalysis;
     using FeatureToggles;
     using FeatureToggles.Configuration;
+    using FeatureToggles.Configuration.AppConfig.Providers;
     using FeatureToggles.Providers;
     using Moq;
     using NUnit.Framework;
